Fix: 헤더, 사이드바 분리 / 관리자 기능 url 접근권한 설정
