+++ vms/src/main/java/com/example/vms/vacation/controller/VacationController.java
@@ -407,78 +407,4 @@ public class VacationController {
 		return new ResponseEntity<byte[]>(file.getFileData(), headers, HttpStatus.OK);
 	}
 
-	// 모든 사원 일정 조회(관리자)
-	@GetMapping("/schedule/list")
-	public String getSchedule(HttpServletRequest request, Model model) {
-
-		// 쿠키 정보
-		Cookie[] cookies = request.getCookies();
-		String token = "";
-		for (Cookie cookie : cookies) {
-			if (cookie.getName().equals("X-AUTH-TOKEN")) {
-				token = cookie.getValue();
-			}
-		}
-		// 토큰 유효성 검사
-		if (tokenProvider.validateToken(token)) {
-			
-			String empId = tokenProvider.getEmpId(token);
-	        Employee employee = employeeService.selectEmployee(empId);
-	        model.addAttribute("employee", employee);
-			
-			// 관리자인지 확인
-			Authentication auths = tokenProvider.getAuthentication(token);
-			Collection<? extends GrantedAuthority> authorities = auths.getAuthorities();
-			for (GrantedAuthority authority : authorities) {
-				//System.out.println(authority.getAuthority());
-				// String authorityName = authority.getAuthority();
-				if (authority.getAuthority().equals("ROLE_LEADER")) {
-					return "manager/schedulelist";
-				}
-			}
-			return "redirect:/employee/login";
-
-		} else {
-			return "redirect:/employee/login";
-		}
-	}
-
-	@ResponseBody
-	@GetMapping("/schedule/getrow")
-	public int getScheduleRowNum(@RequestParam(name = "keyword", required = false, defaultValue = "") String keyword,
-			@RequestParam(name = "option", required = false, defaultValue = "0") String option) {
-
-		int rowNum = 0;
-
-		// 쿠키 정보
-//	        Cookie[] cookies = request.getCookies();
-//	        String token = "";
-//	        for(Cookie cookie : cookies) {
-//	           if(cookie.getName().equals("X-AUTH-TOKEN")) {
-//	              token = cookie.getValue();
-//	           }
-//	        }
-
-		// 토큰에서 empId 추출
-		// String empId = tokenProvider.getEmpId(token);
-
-		rowNum = vacationService.getCountScheduleByOption(keyword, Integer.parseInt(option));
-
-		return rowNum;
-	}
-
-	@ResponseBody
-	@GetMapping("/schedule/getlist")
-	public List<ScheduleEmpDeptType> getScheduleList(
-			@RequestParam(name = "keyword", required = false, defaultValue = "") String keyword,
-			@RequestParam(name = "option", required = false, defaultValue = "0") String option,
-			@RequestParam(name = "curPage", required = false, defaultValue = "1") String curPage) {
-
-		List<ScheduleEmpDeptType> schedulList = vacationService.getScheduleListByOption(curPage, keyword,
-				Integer.parseInt(option));
-
-		System.out.println("schedule" + schedulList);
-		return schedulList;
-	}
-
 }
