+++ vms/src/main/resources/templates/include/header.html
@@ -1,0 +1,25 @@
+<header class="app-header">
+        <nav class="navbar navbar-expand-lg navbar-light">
+          <ul class="navbar-nav">
+            <li class="nav-item d-block d-xl-none">
+              <a class="nav-link sidebartoggler nav-icon-hover" id="headerCollapse" href="javascript:void(0)">
+                <i class="ti ti-menu-2"></i>
+              </a>
+            </li>
+          </ul>
+          <div class="navbar-collapse justify-content-end px-0" id="navbarNav">
+            <ul class="navbar-nav flex-row ms-auto align-items-center justify-content-end">
+				<li class="nav-item" style="margin-right: 15px;">
+					<p th:text="${employee.empId}">사원 아이디</p>
+				</li>
+				<li class="nav-item" style="margin-right: 15px;">
+					<p th:text="${employee.name}">이름</p>
+				</li>
+				<li class="nav-item">
+					<p th:text="${employee.email}">이메일</p>
+				</li>
+            </ul>
+          </div>
+        </nav>
+      </header>
+      <!--  Header End -->
